Add Gradle build. Put packages to java build standards.Format to Google standards. Allow support for mvn install so the Jar can be found locally
