+++ build.gradle
@@ -1,0 +1,31 @@
+group = 'ccperkins'
+version = '0.0.1-SNAPSHOT'
+//https://github.com/ccperkins/FeSimpleGeoProx
+apply plugin: 'java'
+apply plugin: "checkstyle"
+apply plugin: "maven"
+
+sourceCompatibility = '1.8'
+targetCompatibility = '1.8'
+
+version = '0.0.1-SNAPSHOT'
+
+repositories {
+    mavenLocal()
+    mavenCentral()
+}
+
+repositories {
+    mavenLocal()
+    mavenCentral()
+    jcenter()
+}
+
+
+
+dependencies {
+    implementation 'com.javadocmd:simplelatlng:1.3.1'
+    testImplementation(
+            'junit:junit:4.12'
+    )
+}
